# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API for Hamster Money Service Application",
        "title": "Hamster API",
        "contact": {
            "name": "Hamster API Support",
            "url": "https://t.me/CodeMaster482",
            "email": "grigorikovalenko@gmail.com"
        },
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/user/{userID}/account/feed",
    "paths": {
        "/api/account/create": {
            "post": {
                "description": "Create account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Create account",
                "parameters": [
                    {
                        "description": "Input account create",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create account",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_AccountCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/account/update": {
            "put": {
                "description": "Put \taccount",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "PUT \tUpdate",
                "parameters": [
                    {
                        "description": "Input transactin update",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update account",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/account/{account_id}/delete": {
            "delete": {
                "description": "Delete account with chosen ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Delete Account",
                "responses": {
                    "200": {
                        "description": "Account deleted",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Account error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "User unathorized",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "User hasn't rights",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/auth/checkAuth": {
            "post": {
                "description": "Validate auth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Validate Auth",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User status",
                        "schema": {
                            "$ref": "#/definitions/http.Response-auth_SignResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid cookie",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error: cookie read fail",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/auth/signin": {
            "post": {
                "description": "Login account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign In",
                "parameters": [
                    {
                        "description": "username \u0026\u0026 password",
                        "name": "userInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "User logedin",
                        "schema": {
                            "$ref": "#/definitions/http.Response-auth_SignResponse"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "Create Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign Up",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User Created",
                        "schema": {
                            "$ref": "#/definitions/http.Response-auth_SignResponse"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/csrf/": {
            "get": {
                "description": "Get csrf token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csrf"
                ],
                "summary": "Get csrf token",
                "responses": {
                    "200": {
                        "description": "Csrf Token",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_getCSRFResponce"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/tag/all": {
            "get": {
                "description": "Get all tags for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get Tags",
                "responses": {
                    "200": {
                        "description": "tag slice",
                        "schema": {
                            "$ref": "#/definitions/http.Response-array_models_Category"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "auth error relogin",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/tag/create": {
            "post": {
                "description": "Creates tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Create Tag",
                "parameters": [
                    {
                        "description": "tag info",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.TagInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag with id creted",
                        "schema": {
                            "$ref": "#/definitions/http.Response-category_CategoryCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "auth error relogin",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/tag/delete": {
            "delete": {
                "description": "delete tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Delete Tag",
                "parameters": [
                    {
                        "description": "tag id",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag slice",
                        "schema": {
                            "$ref": "#/definitions/http.Response-models_Category"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "auth error relogin",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/tag/{tagId}/update": {
            "put": {
                "description": "Update Tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Update Tag",
                "parameters": [
                    {
                        "description": "tag info",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tag to update",
                        "schema": {
                            "$ref": "#/definitions/http.Response-models_Category"
                        }
                    },
                    "400": {
                        "description": "Incorrect Input",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "auth error relogin",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/count": {
            "get": {
                "description": "Get User count transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get count transaction",
                "responses": {
                    "200": {
                        "description": "Show transaction count",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_TransactionCount"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/create": {
            "post": {
                "description": "Create transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create transaction",
                "parameters": [
                    {
                        "description": "Input transactin create",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_TransactionCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/feed": {
            "get": {
                "description": "Get User all transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get all transaction",
                "parameters": [
                    {
                        "type": "string",
                        "example": "uuid",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "uuid",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-12-21T19:30:57+03:00",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "name": "income",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "name": "outcome",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-11-21T19:30:57+03:00",
                        "name": "start_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_MasTransaction"
                        }
                    },
                    "204": {
                        "description": "Show actual accounts",
                        "schema": {
                            "$ref": "#/definitions/http.Response-string"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/update": {
            "put": {
                "description": "Put transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "PUT Update",
                "parameters": [
                    {
                        "description": "Input transactin update",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CreateAccount": {
            "type": "object",
            "properties": {
                "accumulation": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "number"
                },
                "balance_enabled": {
                    "type": "boolean"
                },
                "mean_payment": {
                    "type": "string"
                }
            }
        },
...
        "http.NilBody": {
            "type": "object"
        },
...
        "http.Response-auth_SignResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/auth.SignResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.Response-http_AccountCreateResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.AccountCreateResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.Response-http_NilBody": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.NilBody"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.UpdateAccount": {
            "type": "object",
            "properties": {
                "accumulation": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "number"
                },
                "balance_enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "mean_payment": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "planned_budget": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}